Добавлена валидация, убраны ненужные переменные, переименованы одноименные переменные, убраны лишние catch'и,

+++ src/main/java/ru/kolodkin/ConverterService/ConverterServiceController.java
@@ -1,15 +1,13 @@
 package ru.kolodkin.ConverterService;
 
+import jakarta.xml.bind.JAXBException;
 import lombok.extern.log4j.Log4j2;
 import lombok.val;
 import org.springframework.core.io.FileSystemResource;
-import org.springframework.core.io.Resource;
 import org.springframework.http.ContentDisposition;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
-import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import ru.kolodkin.ConverterService.converterTool.factory.ConverterFactory;
@@ -20,27 +18,27 @@ import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @RestController
 @Log4j2
 public class ConverterServiceController {
-    final static private String pathDocumentConvert = "src/main/java/ru/kolodkin/ConverterService/convertDocument/";
+    private static final String pathDocumentConvert = "src/main/java/ru/kolodkin/ConverterService/convertDocument/";
 
     @PostMapping("/upload")
     public @ResponseBody String uploadFile(@RequestParam("file") MultipartFile file) {
         String name = file.getOriginalFilename();
-        if (!file.isEmpty()) {
-            try (BufferedOutputStream stream = new BufferedOutputStream(
-                    new FileOutputStream(pathDocumentConvert + file.getOriginalFilename()));) {
-                byte[] bytes = file.getBytes();
-                stream.write(bytes);
-                return "Вы удачно загрузили " + name + "!";
-            } catch (Exception e) {
-                return "Вам не удалось загрузить " + name + " => " + e.getMessage();
-            }
-        } else {
-            return "Вам не удалось загрузить " + name + " потому что файл пустой.";
+        if (file.isEmpty()) {
+            return "Вам не удалось загрузить " + name + ", потому что файл пустой.";
+        }
+
+        try (BufferedOutputStream stream = new BufferedOutputStream(
+                new FileOutputStream(pathDocumentConvert + file.getOriginalFilename()))) {
+            stream.write(file.getBytes());
+            return "Вы удачно загрузили " + name + "!";
+        } catch (Exception exception) {
+            return "Вам не удалось загрузить " + name + " => " + exception.getMessage();
         }
     }
 
@@ -49,67 +47,76 @@ public class ConverterServiceController {
                                         @RequestParam("firstFile") String firstFile,
                                         @RequestParam("secondFile") String secondFile) {
         if (!Validate.validateArgument(converterType, firstFile, secondFile)) {
-            log.error("Входные аргументы некорректны.");
             return "Входные аргументы некорректны.";
         }
 
-        log.info("С расширениями всё в порядке.");
+        if (Files.notExists(Path.of(pathDocumentConvert + firstFile))) {
+            return "Файл для конвертации не загружен на сервер.";
+        }
 
         try (val inputStream = new FileInputStream(pathDocumentConvert + firstFile);
              val outputStream = new FileOutputStream(pathDocumentConvert + secondFile)) {
 
             ConverterFactory.createConverter(ConverterType.valueOf(converterType))
                     .convert(inputStream, outputStream);
-            log.info("Конвертация прошла успешно.");
-        } catch (FileNotFoundException exception) {
-            log.fatal("Файл не найден... " + exception);
-        } catch (IllegalArgumentException exception) {
-            log.fatal("Неправильный ввод входных данных... ", exception);
+
         } catch (Exception exception) {
-            log.fatal("Непредвиденная ошибка... ", exception);
+            return "Вам не удалось конвертировать файл => " + exception.getMessage();
         }
-        return "Конвертация прошла успешно";
+
+        return "Конвертация прошла успешно.";
     }
 
     @GetMapping("/download")
-    public ResponseEntity<Resource> downloadResult(@RequestParam("nameFile") String nameFile) {
+    public ResponseEntity downloadResult(@RequestParam("nameFile") String nameFile) {
+        if (Files.notExists(Path.of(pathDocumentConvert + nameFile))) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Файл: " + nameFile + " на сервере отсутствует!");
+        }
+
         FileSystemResource resource = new FileSystemResource(pathDocumentConvert + nameFile);
         HttpHeaders headers = new HttpHeaders();
-        ContentDisposition disposition = ContentDisposition
+        headers.setContentDisposition(ContentDisposition
                 .builder("attachment")
-                .filename(resource.getFilename())
-                .build();
-        headers.setContentDisposition(disposition);
+                .filename(Objects.requireNonNull(resource.getFilename()))
+                .build()
+        );
+
         return new ResponseEntity<>(resource, headers, HttpStatus.OK);
     }
 
     @GetMapping("/files")
     public String getAllFile() {
-       return  Arrays.stream(new File(pathDocumentConvert).listFiles())
+        val files = Arrays.stream(Objects.requireNonNull(new File(pathDocumentConvert).listFiles()))
                 .map(File::getName)
-                .collect(Collectors.toList()).toString();
+                .collect(Collectors.toList());
+
+        if (files.size() == 0) {
+            return "Нет файлов.";
+        }
+
+        return files.toString();
     }
 
     @PostMapping("/convert/unique")
-    public ResponseEntity<Resource> uniqueConvert(@RequestParam("convertType") String converterType,
-                                                  @RequestParam("firstFile") MultipartFile firstFile,
-                                                  @RequestParam("secondFile") String secondFile) {
-
+    public ResponseEntity uniqueConvert(@RequestParam("convertType") String converterType,
+                                        @RequestParam("firstFile") MultipartFile firstFile,
+                                        @RequestParam("secondFile") String secondFile) {
         String fileName = firstFile.getOriginalFilename();
         if (!Validate.validateArgument(converterType, fileName, secondFile)) {
-            log.error("Входные аргументы некорректны.");
-            return (ResponseEntity<Resource>) ResponseEntity.status(HttpStatus.CONFLICT);
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Входные аргументы некорректны.");
         }
 
-        log.info("С расширениями всё в порядке.");
-
+        if (firstFile.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(
+                    "Вам не удалось загрузить " + fileName + ", потому что файл пустой.");
+        }
 
         try (BufferedOutputStream stream = new BufferedOutputStream(
                 new FileOutputStream(pathDocumentConvert + fileName));) {
-            byte[] bytes = firstFile.getBytes();
-            stream.write(bytes);
-        } catch (IOException e) {
-            return (ResponseEntity<Resource>) ResponseEntity.status(HttpStatus.CONFLICT);
+            stream.write(firstFile.getBytes());
+        } catch (IOException exception) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Вам не удалось загрузить " + fileName + " => " + exception.getMessage());
         }
 
         try (val inputStream = new FileInputStream(pathDocumentConvert + fileName);
@@ -117,22 +124,22 @@ public class ConverterServiceController {
 
             ConverterFactory.createConverter(ConverterType.valueOf(converterType))
                     .convert(inputStream, outputStream);
-            log.info("Конвертация прошла успешно.");
-        } catch (FileNotFoundException exception) {
-            log.fatal("Файл не найден... " + exception);
-        } catch (IllegalArgumentException exception) {
-            log.fatal("Неправильный ввод входных данных... ", exception);
-        } catch (Exception exception) {
-            log.fatal("Непредвиденная ошибка... ", exception);
+
+        } catch (IOException exception) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Вам не удалось конвертировать файл => " + exception.getMessage());
+        } catch (JAXBException exception) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Ошибка при преобразовании содержания файла в объект. Проверьте правильность ваших данных непосредственно в файле.");
         }
 
         FileSystemResource resource = new FileSystemResource(pathDocumentConvert + secondFile);
         HttpHeaders headers = new HttpHeaders();
         ContentDisposition disposition = ContentDisposition
                 .builder("attachment")
-                .filename(resource.getFilename())
+                .filename(Objects.requireNonNull(resource.getFilename()))
                 .build();
         headers.setContentDisposition(disposition);
-        return new ResponseEntity<>(resource, headers, HttpStatus.OK);
+        return new ResponseEntity(resource, headers, HttpStatus.OK);
     }
 }
